+++ services/sinner/api/doc.go
@@ -1,2 +1,0 @@
-// Package api contains the api used by sinner
-package api