+++ services/sinner/config/doc.go
@@ -1,2 +1,0 @@
-// Package config holds the config for both the server and indexer.
-package config