+++ services/sinner/db/sql/mysql/logger.go
@@ -1,8 +1,0 @@
-package mysql
-
-import (
-	"github.com/ipfs/go-log"
-)
-
-// Logger is the mysql logger.
-var logger = log.Logger("synapse-mysql")