+++ services/sinner/fetcher/client/doc.go
@@ -1,2 +1,0 @@
-// Package client provides a GraphQL client for consuming logs.
-package client