+++ services/sinner/graphql/server/doc.go
@@ -1,2 +1,0 @@
-// Package server provides the GraphQL server.
-package server