+++ services/sinner/graphql/server/graph/resolver/doc.go
@@ -1,2 +1,0 @@
-// Package resolvers provides the GraphQL resolver for the explorer service.
-package resolvers