+++ services/sinner/service/doc.go
@@ -1,2 +1,0 @@
-// Package service is the sinner service.
-package service