+++ services/stiprelayer/relayer/relayer.go
@@ -3,12 +3,10 @@ package relayer
 import (
 	"context"
 	"fmt"
+	"github.com/synapsecns/sanguine/contrib/screener-api/client"
 	"math/big"
 	"strings"
 	"time"
-
-	"go.opentelemetry.io/otel/attribute"
-	"go.opentelemetry.io/otel/trace"
 
 	"github.com/cenkalti/backoff"
 	"github.com/ethereum/go-ethereum/accounts/abi/bind"
@@ -23,6 +21,7 @@ import (
 	"github.com/synapsecns/sanguine/services/stiprelayer/db"
 	"github.com/synapsecns/sanguine/services/stiprelayer/stipapi"
 	"github.com/synapsecns/sanguine/services/stiprelayer/stipconfig"
+	"go.opentelemetry.io/otel/attribute"
 	"golang.org/x/sync/errgroup"
 	"golang.org/x/time/rate"
 )
@@ -42,6 +41,8 @@ type STIPRelayer struct {
 	submittter    submitter.TransactionSubmitter
 	signer        signer.Signer
 	apiServer     *stipapi.Server
+	// screener is used to screen addresses using the screener-api client stub.
+	screener client.ScreenerClient
 }
 
 // NewSTIPRelayer creates a new STIPRelayer with the provided context and configuration.
@@ -62,6 +63,15 @@ func NewSTIPRelayer(ctx context.Context,
 		return nil, fmt.Errorf("could not get api server: %w", err)
 	}
 
+	screener, _ := client.NewNoOpClient()
+
+	if cfg.ScreenerAPIUrl != "" {
+		screener, err = client.NewClient(handler, cfg.ScreenerAPIUrl)
+		if err != nil {
+			return nil, fmt.Errorf("could not create screener client: %w", err)
+		}
+	}
+
 	return &STIPRelayer{
 		cfg:           cfg,
 		db:            store,
@@ -70,6 +80,7 @@ func NewSTIPRelayer(ctx context.Context,
 		submittter:    sm,
 		signer:        sg,
 		apiServer:     apiServer,
+		screener:      screener,
 	}, nil
 }
 
@@ -153,13 +164,9 @@ func (s *STIPRelayer) Run(ctx context.Context) error {
 		return nil
 	})
 
-	err := s.ProcessExecutionResults(ctx, "bridge")
+	err := s.ProcessExecutionResults(ctx)
 	if err != nil {
-		return fmt.Errorf("error processing execution results for bridge: %w", err)
-	}
-	err = s.ProcessExecutionResults(ctx, "rfq")
-	if err != nil {
-		return fmt.Errorf("error processing execution results for rfq: %w", err)
+		return fmt.Errorf("error processing execution results: %w", err)
 	}
 
 	// Start the ticker goroutine for requesting and storing execution results
@@ -203,15 +210,10 @@ func (s *STIPRelayer) RequestAndStoreResults(ctx context.Context) error {
 			//nolint: wrapcheck
 			return ctx.Err() // exit if context is canceled
 		case <-ticker.C:
-			if err := s.ProcessExecutionResults(ctx, "bridge"); err != nil {
+			err := s.ProcessExecutionResults(ctx)
+			if err != nil {
 				// Log the error and decide whether to continue based on the error
-				fmt.Printf("Error processing execution results for bridge: %v", err)
-				// Optionally, you can return the error to stop the goroutine
-				// return err
-			}
-			if err := s.ProcessExecutionResults(ctx, "rfq"); err != nil {
-				// Log the error and decide whether to continue based on the error
-				fmt.Printf("Error processing execution results for rfq: %v", err)
+				fmt.Printf("Error processing execution results: %v", err)
 				// Optionally, you can return the error to stop the goroutine
 				// return err
 			}
@@ -220,18 +222,19 @@ func (s *STIPRelayer) RequestAndStoreResults(ctx context.Context) error {
 }
 
 // ProcessExecutionResults encapsulates the logic for requesting and storing execution results.
-func (s *STIPRelayer) ProcessExecutionResults(parentCtx context.Context, queryType string) (err error) {
+func (s *STIPRelayer) ProcessExecutionResults(parentCtx context.Context) (err error) {
 	fmt.Println("Starting execution logic")
 
-	ctx, span := s.handler.Tracer().Start(parentCtx, "ProcessExecutionResults", trace.WithAttributes(attribute.String("queryType", queryType)))
+	ctx, span := s.handler.Tracer().Start(parentCtx, "ProcessExecutionResults")
 	defer func() {
 		metrics.EndSpanWithErr(span, err)
 	}()
 
-	executionID, err := s.ExecuteDuneQuery(ctx, queryType)
+	executionID, err := s.ExecuteDuneQuery(ctx)
 	if err != nil {
 		return fmt.Errorf("failed to execute Dune query: %w", err)
 	}
+	span.SetAttributes(attribute.String("execution_id", executionID))
 
 	// TODO: remove if exponentialBackoff.InitialInterval waits 30 seconds?
 	// time.Sleep(30 * time.Second) // Consider replacing this with a more robust solution
@@ -262,12 +265,20 @@ func (s *STIPRelayer) ProcessExecutionResults(parentCtx context.Context, queryTy
 	}
 
 	var rowsAfterStartDate []Row
+	var firstResultTime time.Time
 	for _, row := range getResultsJSONResult.Result.Rows {
 		// TODO: Will this panic if StartDate not set?
 		if row.BlockTime.After(s.cfg.StartDate) {
 			rowsAfterStartDate = append(rowsAfterStartDate, row)
 		}
+		if firstResultTime.IsZero() || row.BlockTime.Before(firstResultTime) {
+			firstResultTime = row.BlockTime.Time
+		}
 	}
+	span.SetAttributes(
+		attribute.Int("number_of_rows", len(rowsAfterStartDate)),
+		attribute.String("first_result_time", firstResultTime.String()),
+	)
 	fmt.Println("Number of rows after start date:", len(rowsAfterStartDate))
 
 	// Convert each Row to a STIPTransactions and store them in the database
@@ -366,8 +377,25 @@ func (s *STIPRelayer) RelayAndRebateTransactions(ctx context.Context) error {
 	return nil
 }
 
+const stipRuleset = "stip"
+
 // SubmitAndRebateTransaction handles the relaying and rebating of a single transaction.
+// nolint: cyclop
 func (s *STIPRelayer) SubmitAndRebateTransaction(ctx context.Context, transaction *db.STIPTransactions) error {
+	// Check if the address is blocked
+	blocked, err := s.screener.ScreenAddress(ctx, stipRuleset, transaction.Address)
+	if err != nil {
+		return fmt.Errorf("could not screen address: %w", err)
+	}
+
+	if blocked {
+		err = s.db.UpdateSTIPTransactionDoNotProcess(ctx, transaction.Hash)
+		if err != nil {
+			return fmt.Errorf("could not update STIP transaction as do not process: %w", err)
+		}
+		return fmt.Errorf("address is blocked: %s", transaction.Address)
+	}
+
 	// Calculate the transfer amount based on transaction details
 	// This function encapsulates the logic for determining the transfer amount
 	// You can define it elsewhere and call it here